+++ src/contraction/src/dijkstra_function.cpp
@@ -8,7 +8,7 @@ using namespace std;
 using namespace boost;
 typedef adjacency_list<vecS, vecS, undirectedS, Vertex,Edge> G;
 //calculates the shortest path,returns the pathsize and also the path as an array of edges
-int compute_dijkstra(int64_t source,int64_t target,Edge *edges,int64_t num_vertices,int64_t num_edges,Edge **path)
+int64_t compute_dijkstra(int64_t source,int64_t target,Edge *edges,int64_t num_vertices,int64_t num_edges,Edge **path)
 {
 	int64_t path_size=0;
 	typedef My_dijkstra<G> Graph;
@@ -20,7 +20,7 @@ int compute_dijkstra(int64_t source,int64_t target,Edge *edges,int64_t num_verti
 
 
 
-int dijkstra_on_contracted(int64_t source,int64_t target,
+int64_t dijkstra_on_contracted(int64_t source,int64_t target,
 	pgr_contracted_blob **graphInfo,bool has_rcost,Edge **path)
 {
 	int64_t num_vertices=1;
