+++ src/contraction/src/my_astar_wrapper.cpp
@@ -8,7 +8,7 @@ using namespace std;
 using namespace boost;
 typedef adjacency_list<vecS, vecS, directedS, Vertex,Edge> G;
 //calculates the shortest path,returns the pathsize and also the path as an array of edges
-int compute_astar(int64_t source,int64_t target,Edge *edges,int64_t num_vertices,int64_t num_edges,Edge **path)
+int64_t compute_astar(int64_t source,int64_t target,Edge *edges,int64_t num_vertices,int64_t num_edges,Edge **path)
 {
   int64_t path_size=0;
   typedef My_Astar<G> Graph;
